requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
redis==5.0.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# OpenAI & AI
openai==1.3.7
tiktoken==0.5.2
langfuse==2.7.3
pinecone-client==3.0.3

# Document Processing
# docling==2.0.0  # Optional: Advanced chunking (requires pydantic-settings>=2.3.0)
# transformers==4.36.0  # Optional: For Docling tokenizer support
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Message Queue
pika==1.3.2
aio-pika==9.3.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Logging
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

